+++ packages/modos-static-site/components/index.ts
@@ -2,9 +2,11 @@ export * from './layout/Header';
 export * from './layout/Footer';
 export * from './layout/Layout';
 export * from './layout/ContainerWithMargin';
-export * from './general-ui/Leaflet';
+
 export * from './general-ui/LanguageSelect';
 export * from './general-ui/CardsModos';
 export * from './general-ui/SliderModos';
 
+export * from './map-app/Leaflet';
+export * from './map-app/NavigationPanel';
